Added Coord struct. Changed Point struct component types from int to float. Added deadzone functions. Removed deadzone code from GlfwGamepad class. Added deadzone code to demo.

+++ input_lib/include/input_lib/point.hpp
@@ -4,10 +4,10 @@ namespace InputLib
 {
 struct Point
 {
-    int x;
-    int y;
+    float x;
+    float y;
 
-    Point(int x = 0, int y = 0)
+    Point(float x = 0, float y = 0)
     : x(x), y(y) {
     }
 };
